+++ CHANGES.rst
@@ -4,8 +4,8 @@
 New Features
 ------------
 
-- There are now `show_in_sidecar ` and `show_in_new_tab` methods on all the
-  helpers that display the viewers in separate jupyterlab windows from the 
+- There are now ``show_in_sidecar` ` and ``show_in_new_tab`` methods on all the
+  helpers that display the viewers in separate JupyterLab windows from the 
   notebook. [#952]
 
 Cubeviz